Database seeding script works properly
