+++ package.json
@@ -40,7 +40,7 @@
     "test": "jest",
     "react-dev": "webpack -d --watch",
     "server-dev": "nodemon server/index.js",
-    "seed": "node database/faker.js"
+    "seed": "node database/db.js"
   },
   "repository": {
     "type": "git",
